requirements.txt
numpy
pandas
scikit-learn
joblib
pyyaml
dvc
dvc-gdrive
mlflow
matplotlib
seaborn
